Split employee and client forms. Add resend email functionality.
